fix(test): decoding emojis is broken
The `call#receiveText` doesn't decode the emojis in the body content.

The solution is to use `call#receive<ByteArray>` to get the bytes array of the body content.
Then, construct new `String` object and pass the bytes array into it.
This way the emojis are being decoded in a proper way.